Modularizes the fake sysfs and devfs generation logic into a reusable library.
